Begin removing unused code
