+++ families.py
@@ -131,9 +131,6 @@ def build_FAMILIES():
         (["G1", "G2", "G3", "G4", "G6", "G8"], [n%2==1, m%3==1,6*n<=5*m,2*m+4<3*n]),
         (["G1", "G2", "G3", "G4", "G6", "G9"], [n%2==1, m%3==2,6*n<=5*m,2*m+4<3*n]),
 
-        # (["G1", "G2", "G3", "G6"], [m%3==0,3*n<=2*m+4]),
-        # (["G1", "G2", "G3", "G7b", "G8"], [m%3==1,3*n<=2*m+4]),
-        # (["G1", "G2", "G3", "G6", "G9"], [m%3==2,3*n<=2*m+4]),
         (["G1", "G2", "G3", "G6"], [m%3==0,3*n/2-2<=m]),
         (["G1", "G2", "G3", "G7b", "G8"], [m%3==1,3*n/2-2<=m]),
         (["G1", "G2", "G3", "G6", "G9"], [m%3==2,3*n/2-2<=m]),
